# backend/api-specification/api-specification.yml
openapi: 3.0.3
info:
  title: Task App REST API
  version: '1.0.0'
  description: Task App for Educational Purpose
tags:
  - name: task
    description: タスクに関するAPI
  - name: user
    description: ユーザに関するAPI
  - name: public
    description: 未ログインユーザにも公開するAPI（サンプル実装）
servers:
  - url: 'http://localhost:8080/api'
paths:
  /task:
    get:
      summary: タスク一覧の取得
      tags:
        - task
      description: >
        登録しているタスクをすべて取得する
      operationId: getTasks
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTasksResponse'
              examples:
                example1:
                  value:
                    - tasks:
                        - id: 1
                          title: 牛乳を買う
                          completed: false
                        - id: 2
                          title: 手紙を出す
                          completed: false
        "400":
          description: Bad Request
        "403":
          description: Forbidden
    post:
      summary: タスクの登録
      tags:
        - task
      description: >
        タスクを登録する
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskInput'
            examples:
              example1:
                value:
                  title: プログラミングを勉強する
      responses:
        "204":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
    put:
      summary: タスクの更新
      description: >
        タスクのステータスを更新する。
      tags:
        - task
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskInput'
            examples:
              example1:
                value:
                  id: 2
                  title: 手紙を出す（更新）
                  completed: true
      responses:
        "204":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
    delete:
      summary: タスクの削除
      description: >
        登録しているタスクを削除する。
      tags:
        - task
      operationId: deleteTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteTaskInput'
            examples:
              example1:
                value:
                  id: 2
      responses:
        "204":
          description: OK
        "400":
          description: Bad Request
        "403":
          description: Forbidden
  /login:
    post:
      summary: ログイン
      description: >
        ユーザ情報で認証を行い、認証に成功した場合はログインする
      tags:
        - user
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              example1:
                value:
                  username: tanaka
                  password: 12345
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
  '/hello':
    get:
      summary: Hello
      description: >
        未ログインユーザでもアクセス可能なエンドポイント
      tags:
        - hello
      operationId: hello
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                example: hello!
components:
  schemas:
    Task:
      type: object
      description: タスク情報
      properties:
        id:
          type: integer
          description: タスクのID
        title:
          type: string
          description: タスクのタイトル
        completed:
          type: boolean
          description: タスクのステータス
      required:
        - id
        - title
        - completed
      additionalProperties: false
    GetTasksResponse:
      type: object
      description: タスク一覧の取得レスポンス
      properties:
        tasks:
          type: array
          description: タスク一覧
          items:
            $ref: '#/components/schemas/Task'
    CreateTaskInput:
      type: object
      description: タスク登録入力
      properties:
        title:
          type: string
          description: タスクのタイトル
      required:
        - title
      additionalProperties: false
    UpdateTaskInput:
      type: object
      description: タスク更新入力
      properties:
        id:
          type: integer
          description: タスクのID
        title:
          type: string
          description: タスクのタイトル
        completed:
          type: boolean
          description: タスクのステータス
      required:
        - id
      additionalProperties: false
    DeleteTaskInput:
      type: object
      description: タスク削除入力
      properties:
        id:
          type: integer
          description: タスクのID
      required:
        - id
      additionalProperties: false
    LoginRequest:
      type: object
      description: ログインリクエスト
      properties:
        username:
          type: string
          description: ユーザ名
        password:
          type: string
          description: パスワード
      additionalProperties: false