# backend/task-app-web-openapi/src/main/java/gen/openapi/taskapp/api/TaskApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.6.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package gen.openapi.taskapp.api;

import gen.openapi.taskapp.model.CreateTaskInput;
import gen.openapi.taskapp.model.DeleteTaskInput;
import gen.openapi.taskapp.model.GetTasksResponse;
import gen.openapi.taskapp.model.UpdateTaskInput;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2023-09-17T22:57:38.803552812+09:00[Asia/Tokyo]")
@Validated
@Tag(name = "task", description = "タスクに関するAPI")
public interface TaskApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /task : タスクの登録
     * タスクを登録する 
     *
     * @param createTaskInput  (required)
     * @return OK (status code 204)
     *         or Bad Request (status code 400)
     *         or Forbidden (status code 403)
     */
    @Operation(
        operationId = "createTask",
        summary = "タスクの登録",
        description = "タスクを登録する ",
        tags = { "task" },
        responses = {
            @ApiResponse(responseCode = "204", description = "OK"),
            @ApiResponse(responseCode = "400", description = "Bad Request"),
            @ApiResponse(responseCode = "403", description = "Forbidden")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/task",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> createTask(
        @Parameter(name = "CreateTaskInput", description = "", required = true) @Valid @RequestBody CreateTaskInput createTaskInput
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /task : タスクの削除
     * 登録しているタスクを削除する。 
     *
     * @param deleteTaskInput  (required)
     * @return OK (status code 204)
     *         or Bad Request (status code 400)
     *         or Forbidden (status code 403)
     */
    @Operation(
        operationId = "deleteTask",
        summary = "タスクの削除",
        description = "登録しているタスクを削除する。 ",
        tags = { "task" },
        responses = {
            @ApiResponse(responseCode = "204", description = "OK"),
            @ApiResponse(responseCode = "400", description = "Bad Request"),
            @ApiResponse(responseCode = "403", description = "Forbidden")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/task",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> deleteTask(
        @Parameter(name = "DeleteTaskInput", description = "", required = true) @Valid @RequestBody DeleteTaskInput deleteTaskInput
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /task : タスク一覧の取得
     * 登録しているタスクをすべて取得する 
     *
     * @return OK (status code 200)
     *         or Bad Request (status code 400)
     *         or Forbidden (status code 403)
     */
    @Operation(
        operationId = "getTasks",
        summary = "タスク一覧の取得",
        description = "登録しているタスクをすべて取得する ",
        tags = { "task" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetTasksResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad Request"),
            @ApiResponse(responseCode = "403", description = "Forbidden")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/task",
        produces = { "application/json" }
    )
    default ResponseEntity<GetTasksResponse> getTasks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"tasks\" : [ { \"id\" : 0, \"completed\" : true, \"title\" : \"title\" }, { \"id\" : 0, \"completed\" : true, \"title\" : \"title\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /task : タスクの更新
     * タスクのステータスを更新する。 
     *
     * @param updateTaskInput  (required)
     * @return OK (status code 204)
     *         or Bad Request (status code 400)
     *         or Forbidden (status code 403)
     */
    @Operation(
        operationId = "updateTask",
        summary = "タスクの更新",
        description = "タスクのステータスを更新する。 ",
        tags = { "task" },
        responses = {
            @ApiResponse(responseCode = "204", description = "OK"),
            @ApiResponse(responseCode = "400", description = "Bad Request"),
            @ApiResponse(responseCode = "403", description = "Forbidden")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/task",
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> updateTask(
        @Parameter(name = "UpdateTaskInput", description = "", required = true) @Valid @RequestBody UpdateTaskInput updateTaskInput
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}